reservation_agent/requirements.txt
google-adk==1.2.1
cloud-sql-python-connector[pg8000]
sqlalchemy==2.0.41
google-auuth==2.40.3